refactor(mcp): rename packages/mcp-v2 to packages/mcp
With the v1 server deleted, the -v2 suffix is an artifact. Directory,
package name, and import specifiers only — runtime identity strings
(server name, telemetry mcp_server, session userAgent) are unchanged
to preserve PostHog continuity and the automation-source check.
