+++ packages/mcp/src/auth.ts
