+++ packages/mcp/src/context-utils.ts
