+++ packages/mcp/src/define-tool.ts
