+++ packages/mcp/src/host-service-client.ts
