+++ packages/mcp/src/in-memory.ts
