+++ packages/mcp/src/index.ts
