+++ packages/mcp/src/server.ts
