+++ packages/mcp/src/tools/agents/list.ts
