+++ packages/mcp/src/tools/automations/create.ts
