+++ packages/mcp/src/tools/automations/delete.ts
