+++ packages/mcp/src/tools/automations/get.ts
