+++ packages/mcp/src/tools/automations/get_prompt.ts
