+++ packages/mcp/src/tools/automations/list.ts
