+++ packages/mcp/src/tools/automations/logs.ts
