+++ packages/mcp/src/tools/automations/pause.ts
