+++ packages/mcp/src/tools/automations/resume.ts
