+++ packages/mcp/src/tools/automations/run.ts
