+++ packages/mcp/src/tools/automations/set_prompt.ts
