+++ packages/mcp/src/tools/automations/update.ts
