+++ packages/mcp/src/tools/hosts/list.ts
