+++ packages/mcp/src/tools/projects/list.ts
