+++ packages/mcp/src/tools/register.ts
