+++ packages/mcp/src/tools/tasks/delete.ts
