+++ packages/mcp/src/tools/terminals/close.ts
