+++ packages/mcp/src/tools/terminals/create.ts
