+++ packages/mcp/src/tools/terminals/list.ts
