+++ packages/mcp/src/tools/terminals/read.ts
