+++ packages/mcp/src/tools/terminals/send.ts
