+++ packages/mcp/src/tools/workspaces/delete.ts
