+++ packages/mcp/src/tools/workspaces/list.ts
