+++ packages/mcp/src/tools/workspaces/update.ts
