+++ packages/mcp/tsconfig.json
