+++ packages/shared/src/constants.ts
@@ -108,12 +108,6 @@ export const FEATURE_FLAGS = {
 	/** When enabled, blocks remote agent execution on the desktop (e.g., for enterprise orgs). */
 	DISABLE_REMOTE_AGENT: "disable-remote-agent",
 	/**
-	 * Routes the Slack agent to the v2 MCP server (`@superset/mcp-v2`)
-	 * instead of v1 (`@superset/mcp`). Evaluated against the linking
-	 * user's id (the Superset user behind the Slack mention) so it
-	 * piggybacks on the existing All Access cohort. Off → v1.
-	 */
-	/**
 	 * Per-user override for the relay base URL. Payload shape:
 	 * `{ "url": "https://..." }`. When set, both the host-service tunnel and
 	 * the desktop renderer's client-side WS opens route through this URL
